pyproject.toml
[build-system]
requires = ["hatchling>=1.21.1", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "request_ddi"
description = "Discover, compare and reuse questions from social sciences and humanities surveys"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.9"
authors = [
  { name = "William Feraux", email = "william.feraux@sciencespo.fr" },
  { name = "Malaury Lemaitre-Salmon", email = "malaury.lemaitresalmon@sciencespo.fr" },
]
keywords = [
  "django",
  "elasticsearch",
  "surveys",
  "ddi",
  "documentation",
  "metadata",
  "socialsciences",
]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: Django",
  "Framework :: Django :: 5",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]

dependencies = [
  "elasticsearch==9.2.0",
  "django-elasticsearch-dsl==9.0.0",
  "django-ninja==1.3.0",
  "beautifulsoup4",
  "lxml",
  "django-import-export==4.1.1",
  "psycopg2-binary",
  "gunicorn==23.0.0",
  "whitenoise==6.4.0",
  "django-extensions>=3.3",
  "django-health-check",
  "python-dotenv",
  "requests",
]
dynamic = [
  "version",
]

[project.scripts]
request_ddi_manage = "request_ddi.manage:main"

[project.urls]
Homepage = "https://gitlab.sciences-po.fr/cdspit/request/request-ddi"
Documentation = "http://cdspit.gitlab-pages.sciences-po.fr/request/docs/"
Source = "https://gitlab.sciences-po.fr/cdspit/request/request-ddi"
Issues = "https://gitlab.sciences-po.fr/cdspit/request/request-ddi/-/issues"

[project.optional-dependencies]
test = [
  "coverage",
  "pytest>=7.0",
  "pytest-django",
  "pytest-cov",
]

# Check ruff version is aligned with the one in .pre-commit-config.yaml
dev = [
  "request_ddi[test]",
  "build",
  "pre-commit",
  "pytest-cov",
  "coverage",
  "hatch",
  "ruff",
  "gitlab_activity",
]

[tool.check-wheel-contents]
ignore = ["W002", "W004"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "request_ddi/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["request_ddi"]

[tool.hatch.build.targets.wheel.shared-data]
"share/locale" = "share/request_ddi/locale"

[tool.pytest]
DJANGO_SETTINGS_MODULE = "config.test_settings"
pythonpath = [".", "config"]
python_files = ["request_ddi/tests/test_*.py"]

[tool.ruff]
exclude = ["*.ipynb", "request_ddi/_version.py"]
target-version = "py39"
line-length = 100
lint.select = [
  "A", "B", "C", "DTZ", "E", "EM", "F", "FBT", "I", "ICN", "N",
  "PLC", "PLE", "PLR", "PLW", "Q", "RUF", "S", "SIM", "T", "TID", "UP",
  "W", "YTT",
]
lint.ignore = [
	# FBT001 Boolean positional arg in function definition
	"FBT001", "FBT002", "FBT003",
	# PLR0913 Too many arguments to function call
	"PLR0913",
	# SIM105 Use `contextlib.suppress(...)`
	"SIM105",
	# E501 Line too long (112 > 100)
	"E501",
]
# In addition to the standard set of exclusions, omit all tests, plus a specific file.
extend-exclude = ["request_ddi/migrations/*"]

[tool.ruff.lint.per-file-ignores]
# E501 Line too long (158 > 100 characters)
"request_ddi/tests/*" = ["E501"]
# RUF012 Mutable class attributes should be annotated with `typing.ClassVar`
"request_ddi/migrations/*" = ["RUF012"]
